DIABLO-720, all email-related jobs consolidated into one 'emails' job
