+++ diablo/jobs/tasks/invitation_emails_task.py
@@ -22,42 +22,31 @@ SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED HEREUNDER IS PROVIDED
 "AS IS". REGENTS HAS NO OBLIGATION TO PROVIDE MAINTENANCE, SUPPORT, UPDATES,
 ENHANCEMENTS, OR MODIFICATIONS.
 """
-import os
-
-from diablo import cache, db
-from diablo.configs import load_configs
-from diablo.jobs.background_job_manager import BackgroundJobManager
-from diablo.logger import initialize_logger
-from diablo.routes import register_routes
-from flask import Flask
-
-background_job_manager = BackgroundJobManager()
+from diablo.jobs.tasks.base_task import BaseTask
+from diablo.models.email_template import EmailTemplate
+from diablo.models.queued_email import QueuedEmail
+from diablo.models.sis_section import SisSection
+from flask import current_app as app
 
 
-def create_app(standalone=False):
-    """Initialize app with configs."""
-    app = Flask(__name__.split('.')[0])
-    load_configs(app)
-    initialize_logger(app)
-    cache.init_app(app)
-    cache.clear()
-    db.init_app(app)
+class InvitationEmailsTask(BaseTask):
 
-    if not standalone:
-        with app.app_context():
-            register_routes(app)
-            _register_jobs(app)
+    def _run(self):
+        self.term_id = app.config['CURRENT_TERM_ID']
+        self.email_new_invites()
 
-    return app
+    @classmethod
+    def description(cls):
+        return f"Queues up '{EmailTemplate.get_template_type_options()['invitation']}' emails."
 
-
-def _register_jobs(app):
-    from diablo.jobs.blackouts_job import BlackoutsJob  # noqa
-    from diablo.jobs.canvas_job import CanvasJob  # noqa
-    from diablo.jobs.house_keeping_job import HouseKeepingJob  # noqa
-    from diablo.jobs.kaltura_job import KalturaJob  # noqa
-    from diablo.jobs.emails_job import EmailsJob  # noqa
-    from diablo.jobs.sis_data_refresh_job import SisDataRefreshJob  # noqa
-
-    if app.config['JOBS_AUTO_START'] and (not app.debug or os.environ.get('WERKZEUG_RUN_MAIN') == 'true'):
-        background_job_manager.start(app)
+    def email_new_invites(self):
+        for course in SisSection.get_courses(term_id=self.term_id):
+            if not course['hasOptedOut'] and len(course.get('meetings', {}).get('eligible', [])) == 1:
+                for i in course['instructors']:
+                    if not i['wasSentInvite']:
+                        QueuedEmail.create(
+                            recipient=i,
+                            section_id=course['sectionId'],
+                            template_type='invitation',
+                            term_id=self.term_id,
+                        )
